# Conflicts:
#	src/main/java/tfc/dynamicportals/mixin/client/quality/ClientChunkCacheMixin.java
#	src/main/java/tfc/dynamicportals/mixin/common/quality/ChunkMapMixin.java

+++ src/main/java/tfc/dynamicportals/GLUtils.java
@@ -24,7 +24,7 @@ public class GLUtils {
 		ShaderInstance shaderinstance = GameRenderer.getPositionColorShader();
 		Matrix4f project = RenderSystem.getProjectionMatrix();
 
-		Matrix4f matrix4f = Matrix4f.orthographic((float) pWidth, (float) (-pHeight), 1000.0F, 3000.0F);
+		Matrix4f matrix4f = Matrix4f.orthographic(pWidth, -pHeight, 1000.0F, 3000.0F);
 		RenderSystem.setProjectionMatrix(matrix4f);
 		if (shaderinstance.MODEL_VIEW_MATRIX != null) {
 			shaderinstance.MODEL_VIEW_MATRIX.set(Matrix4f.createTranslateMatrix(0.0F, 0.0F, -2000.0F));
